helpers: factor out date_range_list()
